requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
pypdf
bs4
arxiv
pymupdf
langchain-text-splitters
langchain.openai
faiss-cpu
chromadb
langchain_chroma
langchain-groq
streamlit
transformers
openai-whisper
ffmpeg-python
tiktoken
langchain-embeddings
langchain-llms